refactor: move upload_to_catbox to commons.functions
Per PR #26 review feedback (monkut). The catbox uploader is a
generic public-URL helper, not Instagram-specific — relocates the
helper, its CATBOX_API_URL constant, and CatboxUploadError to
commons/functions.py so other consumers can reuse it.

instagram_post.py now imports upload_to_catbox from commons.functions.
The test for the helper moves to commons/tests/test_functions.py;
tests for the Instagram-specific flow stay in test_instagram_post.py.

No runtime behaviour change.
